Decouple big dialog bindings file. Testing of dialog callbacks.
