+++ src/engine/scripts/declarations/dialogs/dialogs/index.ts
@@ -1,0 +1,2 @@
+require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_generic");
+require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_quest");
