+++ src/engine/scripts/declarations/dialogs/dialogs_zaton.ts
@@ -46,8 +46,6 @@ import {
 
 const logger: LuaLogger = new LuaLogger($filename);
 
-extern("dialogs_zaton", {});
-
 /**
  * todo;
  */
