+++ src/engine/scripts/declarations/dialogs/index.ts
@@ -1,6 +1,6 @@
+require("@/engine/scripts/declarations/dialogs/dialogs");
 require("@/engine/scripts/declarations/dialogs/dialogs_pripyat");
 require("@/engine/scripts/declarations/dialogs/dialogs_jupiter");
 require("@/engine/scripts/declarations/dialogs/dialogs_zaton");
 
-require("@/engine/scripts/declarations/dialogs/dialogs");
 require("@/engine/scripts/declarations/dialogs/dialog_manager");
